# openapi.yaml
openapi: 3.0.3
info:
  title: ManadaBook API
  description: API completa para la red social de mascotas más grande del mundo
  version: 1.0.0
  contact:
    name: ManadaBook Support
    email: support@manadabook.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.manadabook.com
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Circle:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 100
        slug:
          type: string
          maxLength: 50
          pattern: '^[a-z0-9-]+$'
        type:
          type: string
          enum: [public, private]
        admin_id:
          type: string
          format: uuid
        description:
          type: string
          maxLength: 500
        location:
          type: string
          maxLength: 200
        members_count:
          type: integer
          minimum: 0
        created_at:
          type: string
          format: date-time
      required: [id, name, slug, type, admin_id]
...
paths:
...
  /circles:
    get:
      tags:
        - Circles
      summary: Listar grupos
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
        - name: type
          in: query
          schema:
            type: string
            enum: [public, private]
        - name: location
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Lista de grupos
          content:
            application/json:
              schema:
                type: object
                properties:
                  circles:
                    type: array
                    items:
                      $ref: '#/components/schemas/Circle'
                  total:
                    type: integer
                  page:
                    type: integer
                  hasMore:
                    type: boolean

    post:
      tags:
        - Circles
      summary: Crear nuevo grupo
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 100
                type:
                  type: string
                  enum: [public, private]
                description:
                  type: string
                  maxLength: 500
                location:
                  type: string
                  maxLength: 200
              required: [name, type]
      responses:
        '201':
          description: Grupo creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Circle'
...
  /moderation/actions:
    post:
      tags:
        - Moderation
      summary: Ejecutar acción de moderación
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                target_type:
                  type: string
                  enum: [post, user, message, snippet, ad]
                target_id:
                  type: string
                  format: uuid
                action_type:
                  type: string
                  enum: [warning, content_removal, temporary_ban, permanent_ban, unban]
                reason:
                  type: string
                  maxLength: 500
                duration_days:
                  type: integer
                  minimum: 1
              required: [target_type, target_id, action_type, reason]
      responses:
        '200':
          description: Acción ejecutada

  # Analytics
...